Move some (partial) classes to common
