+++ src/Hl7.Fhir.Core/FhirPath/FhirPathExtensions.cs
@@ -1,0 +1,57 @@
+﻿/* 
+ * Copyright (c) 2016, Firely (info@fire.ly) and contributors
+ * See the file CONTRIBUTORS for details.
+ * 
+ * This file is licensed under the BSD 3-Clause license
+ * available at https://raw.githubusercontent.com/FirelyTeam/firely-net-sdk/master/LICENSE
+ */
+
+using Hl7.Fhir.ElementModel;
+using Hl7.Fhir.Model;
+using Hl7.Fhir.Utility;
+using Hl7.FhirPath;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace Hl7.Fhir.FhirPath
+{
+    public static class FhirPathExtensions
+    {
+        public static Func<string, ITypedElement> ToFhirPathResolver(this Func<string, Resource> resolver)
+        {
+            return navResolver;
+
+            ITypedElement navResolver(string url)
+            {
+                var resource = resolver(url);
+                return resource?.ToTypedElement();
+            }
+        }
+
+        public static IEnumerable<Base> Select(this Base input, string expression, FhirEvaluationContext ctx = null)
+        {
+            var inputNav = input.ToTypedElement();
+            var result = inputNav.Select(expression, ctx ?? FhirEvaluationContext.CreateDefault());
+            return result.ToFhirValues();
+        }
+
+        public static object Scalar(this Base input, string expression, FhirEvaluationContext ctx = null)
+        {
+            var inputNav = input.ToTypedElement();
+            return inputNav.Scalar(expression, ctx ?? FhirEvaluationContext.CreateDefault());
+        }
+
+        public static bool Predicate(this Base input, string expression, FhirEvaluationContext ctx = null)
+        {
+            var inputNav = input.ToTypedElement();
+            return inputNav.Predicate(expression, ctx ?? FhirEvaluationContext.CreateDefault());
+        }
+
+        public static bool IsBoolean(this Base input, string expression, bool value, FhirEvaluationContext ctx = null)
+        {
+            var inputNav = input.ToTypedElement();
+            return inputNav.IsBoolean(expression, value, ctx ?? FhirEvaluationContext.CreateDefault());
+        }
+    }
+}
